move validator to helpers
