+++ cdc/api/v2/api_helpers_mock.go
@@ -40,21 +40,6 @@ func (m *MockAPIV2Helpers) EXPECT() *MockAPIV2HelpersMockRecorder {
 	return m.recorder
 }
 
-// getKVTiStore mocks base method.
-func (m *MockAPIV2Helpers) getKVTiStore(arg0 []string, arg1 *security.Credential) (kv.Storage, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "getKVTiStore", arg0, arg1)
-	ret0, _ := ret[0].(kv.Storage)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// getKVTiStore indicates an expected call of getKVTiStore.
-func (mr *MockAPIV2HelpersMockRecorder) getKVTiStore(arg0, arg1 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "getKVTiStore", reflect.TypeOf((*MockAPIV2Helpers)(nil).getKVTiStore), arg0, arg1)
-}
-
 // getPDClient mocks base method.
 func (m *MockAPIV2Helpers) getPDClient(arg0 context.Context, arg1 []string, arg2 *security.Credential) (client.Client, error) {
 	m.ctrl.T.Helper()
@@ -68,6 +53,21 @@ func (m *MockAPIV2Helpers) getPDClient(arg0 context.Context, arg1 []string, arg2
 func (mr *MockAPIV2HelpersMockRecorder) getPDClient(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "getPDClient", reflect.TypeOf((*MockAPIV2Helpers)(nil).getPDClient), arg0, arg1, arg2)
+}
+
+// getTiStore mocks base method.
+func (m *MockAPIV2Helpers) getTiStore(arg0 []string, arg1 *security.Credential) (kv.Storage, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "getTiStore", arg0, arg1)
+	ret0, _ := ret[0].(kv.Storage)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// getTiStore indicates an expected call of getTiStore.
+func (mr *MockAPIV2HelpersMockRecorder) getTiStore(arg0, arg1 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "getTiStore", reflect.TypeOf((*MockAPIV2Helpers)(nil).getTiStore), arg0, arg1)
 }
 
 // getVerfiedTables mocks base method.
